A typo?
The place in question talks about making gluster believe that the metadata of file `a` on `brick-b` should win. Thus brick `a` should see the metadata changelog of brick `b` to be normal but taking the previous value (`0x000003d70000000100000000`), leave the data log dirty (`000003d7`) and unset the metadata part `00000001`. The `0000000000000000ffffffff` seems to be wrong.