Finally fix rendering of code blocks
Well, the last fix for this (https://github.com/gluster/glusterdocs/pull/210 ) is incorrect, and the problem still persists. After going through the entire documentation repository, found that one other place ("Geo Replication.md") also has the same rendering issue, and that seems to be due to the tab/spaces before the ```

Fixing it right (hopefully) this time.