Guidelines-For-Maintainers: expect good problem description in bugs
When working on a bug, please ensure that you update the BZ with any
relevant information as well as the RCA. I have seen several BZs in the
past which report crashes, however they do not have a bt or RCA
captured.  Having this information in the BZ makes it much easier to see
if a newly reported issue has already been fixed.

Suggested-by: Nithya Balachandran <nbalacha@redhat.com>
URL: https://lists.gluster.org/pipermail/maintainers/2019-April/005850.html
Signed-off-by: Niels de Vos <ndevos@redhat.com>