ctdb requires CTDB_SAMBA_SKIP_SHARE_CHECK to be set
As mentioned at https://www.samba.org/samba/docs/current/man-html/vfs_glusterfs.8.html ctdb checks the path as an absolute path and does not differentiate this for vfs_glusterfs. Therefore ctdb will generate errors if a path is set in the share that does not also exist in the os's file system.
Eg. using `path /` in a samba share config works as it exists in the os files system while `path /blob` might exist in the glusterfs volume but not in the os file system and as such would generate an error