add documentation about hook scripts.
The link [1] on guster.org is is non-existent and Gobinda wanted an upstream
link which can be referenced in the ovirt upstream doc. Hence adding it.

I do not claim originality of the conent. I just copied what is found in
the RHGS admin guidem, which incidentally, seems to be identical to
https://rajeshjoseph.gitbooks.io/test-guide/content/cluster/chap-Managing_Gluster_Volume_Life-Cycle_Extensions.html,
which I am guessing is probably identical to what must have existed in [1].

[1] http://www.gluster.org/community/documentation/index.php/Features/Hooks
Signed-off-by: Ravishankar N <ravishankar@redhat.com>