Adding gfid-mismatch resolution steps
Problem:
Currently gfid split-brains can be healed using the split-brain
resolution CLI or by setting the favorite-child-ploicy to take
care of that automatically. But the documentation still says GFID
split-brains can not be healed automatically and needs to be done
manually.

Fix:
Adding the steps to resolve the gfid mismatch using the split-brain
resolution CLI and updating the other sections appropriately.
The Automagic unsplit-brain section does not need any change as such.

Signed-off-by: karthik-us <ksubrahm@redhat.com>
Fixes: #692