Update client-quorum behaviour for replicate volumes
Problem:
With the fix https://review.gluster.org/#/c/glusterfs/+/17673/ the
client-quorum behaviour has been changed to return ENOTCONN instead
of making the volume EROFS when quorum is lost. But in some places
of the AFR and Arbiter documentations, it is still showing the old
behaviour.

Fix:
Update the documentation to reflect the correct error returned by
the operations when client-quotum is not met.

Signed-off-by: karthik-us <ksubrahm@redhat.com>