[QOL] Hide 'Copy To Clipboard' button on non-code blocks
Sitewide there are multiple code blocks that do not contain
any code in them. But still a `Copy code to clipboard` button
is shown, this patch addresses this issue. It is a QOL improvement
that makes distinguishing code blocks visually easier.

To use this new feature, we need to start our code block like:
```{ .language_name .no-copy }
    ...
    ...
    ...
```

Where language_name is something that can vary (console in our case mostly).
The .no-copy attribute is what disables the function on that particular block.

Signed-off-by: black-dragon74 <niryadav@redhat.com>