[QOL] Extend keyboard shortcuts to copy current address
This patch also removes the obsolete fix-docs.js. That
file depends of jQuery and we no longer use that. That file was also
the cause of the console errors on current docs site.

After this patch, users will be able to press C or Y(ank) to copy
current address along with anchors to the clipboard. This feature is just
an extension of multiple exisiting shortcuts like F/P/N.

Another patch is in the works which would update current address so that
current anchor is in the addressbar. That would make this patch more useful as
you will be able to copy the adress of the location that you are currently reading
in the doc and resume later or maybe share to someone. Cool, right?

Signed-off-by: black-dragon74 <niryadav@redhat.com>