Hide 'Copy to clipboard' on non-code blocks and improvements to `Getting Started` section (#732)
* [QOL] Hide 'Copy To Clipboard' button on non-code blocks

Sitewide there are multiple code blocks that do not contain
any code in them. But still a `Copy code to clipboard` button
is shown, this patch addresses this issue. It is a QOL improvement
that makes distinguishing code blocks visually easier.

To use this new feature, we need to start our code block like:
```{ .language_name .no-copy }
    ...
    ...
    ...
```

Where language_name is something that can vary (console in our case mostly).
The .no-copy attribute is what disables the function on that particular block.

Signed-off-by: black-dragon74 <niryadav@redhat.com>

* [getting-started] Multiple improvements and cleaner syntax

This is a first of many patches that'll follow for subsequent sections sitewide
which moves OLD md syntax to NEW one and also address any other semantics that come
into notice.

Signed-off-by: black-dragon74 <niryadav@redhat.com>

* Properly annotate code blocks

Signed-off-by: black-dragon74 <niryadav@redhat.com>

* Fix code formatter errors

Signed-off-by: black-dragon74 <niryadav@redhat.com>