Update replace-brick documentation (#782)
* docs: update replace-brick notes

We no longer need to do manual steps before replacing a brick. It
is enough to use `replace-brick` without killing brick processes
or setting xattrs manually (this has been true for *years*).

* docs: minor improvements to replace-brick text

Use code blocks for brick name and simplify formatting in places
to be consistent with other examples. Also, it seems that MkDocs
renders this Markdown slightly differently than GitHub.

* docs: reverse order of pure vs distributed replicate

The "pure" replicate example comes first in the docs, so it seems
natural to list it first here.

* Update docs/Administrator-Guide/Managing-Volumes.md

Co-authored-by: Karthik Subrahmanya <ksubrahm@redhat.com>

---------

Co-authored-by: Karthik Subrahmanya <ksubrahm@redhat.com>